make separate test file for fdv2 e2e tests
